Moved config reading/parsing into application.
This makes the processing of the configuration file an explicit action controlled by the user. It also makes it possible for the user to definea configuration by supplying a direct dictionary, or from any other source.

Most importantly, it removes the terribly dirty hack of requiring to poke and parse stack frames.